pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentape"
version = "0.1.0"
description = "Record/replay testing for LLM agents - VCR.py but for LLM SDKs"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "agentape contributors" }
]
keywords = ["llm", "testing", "record", "replay", "openai", "anthropic", "vcr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "pyyaml>=6.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.18.0"]
semantic = ["numpy", "openai"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio",
    "ruff",
    "mypy",
    "types-PyYAML",
]

[project.entry-points.pytest11]
agentape = "agentape.pytest_plugin"

[project.urls]
Homepage = "https://github.com/agentape/agentape"
Documentation = "https://github.com/agentape/agentape#readme"
Repository = "https://github.com/agentape/agentape"
Issues = "https://github.com/agentape/agentape/issues"

[tool.hatch.build.targets.wheel]
packages = ["agentape"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
